input: implement key repeat
In wayland, key repeat is implemented by the client itself. This
requires timing the delay and repeat periods, so implementing a new main
loop was necessary to asynchronously poll the wayland fd and the new
timerfd.

The flush function also now calls wl_display_dispatch to commit frames
on render.